updated middlewares

+++ controllers/admin.js
@@ -1,14 +1,15 @@
 import TryCatch from "../middlewares/TryCatch.js";
 import { Courses } from "../models/Courses.js";
 import { Lecture } from "../models/Lecture.js";
-import { rm } from "fs";
 import { promisify } from "util";
 import fs from "fs";
 import { User } from "../models/User.js";
 
+// Promisify fs.unlink for asynchronous file deletion
+const unlinkAsync = promisify(fs.unlink);
+
 export const createCourse = TryCatch(async (req, res) => {
   const { title, description, category, createdBy, duration, price } = req.body;
-
   const image = req.file;
 
   await Courses.create({
@@ -35,7 +36,6 @@ export const addLectures = TryCatch(async (req, res) => {
     });
 
   const { title, description } = req.body;
-
   const file = req.file;
 
   const lecture = await Lecture.create({
@@ -54,34 +54,40 @@ export const addLectures = TryCatch(async (req, res) => {
 export const deleteLecture = TryCatch(async (req, res) => {
   const lecture = await Lecture.findById(req.params.id);
 
-  rm(lecture.video, () => {
-    console.log("Video deleted");
-  });
+  if (!lecture)
+    return res.status(404).json({
+      message: "No Lecture with this id",
+    });
+
+  await unlinkAsync(lecture.video);
+  console.log("Video deleted");
 
   await lecture.deleteOne();
 
   res.json({ message: "Lecture Deleted" });
 });
 
-const unlinkAsync = promisify(fs.unlink);
-
 export const deleteCourse = TryCatch(async (req, res) => {
   const course = await Courses.findById(req.params.id);
+
+  if (!course)
+    return res.status(404).json({
+      message: "No Course with this id",
+    });
 
   const lectures = await Lecture.find({ course: course._id });
 
   await Promise.all(
     lectures.map(async (lecture) => {
       await unlinkAsync(lecture.video);
-      console.log("video deleted");
+      console.log("Video deleted");
     })
   );
 
-  rm(course.image, () => {
-    console.log("image deleted");
-  });
+  await unlinkAsync(course.image);
+  console.log("Image deleted");
 
-  await Lecture.find({ course: req.params.id }).deleteMany();
+  await Lecture.deleteMany({ course: req.params.id });
 
   await course.deleteOne();
 
@@ -93,12 +99,12 @@ export const deleteCourse = TryCatch(async (req, res) => {
 });
 
 export const getAllStats = TryCatch(async (req, res) => {
-  const totalCoures = (await Courses.find()).length;
+  const totalCourses = (await Courses.find()).length;
   const totalLectures = (await Lecture.find()).length;
   const totalUsers = (await User.find()).length;
 
   const stats = {
-    totalCoures,
+    totalCourses,
     totalLectures,
     totalUsers,
   };
@@ -109,9 +115,7 @@ export const getAllStats = TryCatch(async (req, res) => {
 });
 
 export const getAllUser = TryCatch(async (req, res) => {
-  const users = await User.find({ _id: { $ne: req.user._id } }).select(
-    "-password"
-  );
+  const users = await User.find({ _id: { $ne: req.user._id } }).select("-password");
 
   res.json({ users });
 });
@@ -119,9 +123,15 @@ export const getAllUser = TryCatch(async (req, res) => {
 export const updateRole = TryCatch(async (req, res) => {
   if (req.user.mainrole !== "superadmin")
     return res.status(403).json({
-      message: "This endpoint is assign to superadmin",
+      message: "This endpoint is assigned to superadmin",
     });
+
   const user = await User.findById(req.params.id);
+
+  if (!user)
+    return res.status(404).json({
+      message: "No User with this id",
+    });
 
   if (user.role === "user") {
     user.role = "admin";
@@ -137,7 +147,7 @@ export const updateRole = TryCatch(async (req, res) => {
     await user.save();
 
     return res.status(200).json({
-      message: "Role updated",
+      message: "Role updated to user",
     });
   }
 });
